+++ app/src/main/java/peoples/materialfitness/WorkoutDetails/ExerciseGraph/ExerciseGraphPresenter.java
@@ -108,7 +108,7 @@ public class ExerciseGraphPresenter extends BasePresenter
                 if (exerciseSession.getExercise().equals(exercise))
                 {
                     // Now find the max weight from this exercise session.
-                    int maxWeight = 0;
+                    double maxWeight = 0;
 
                     for (WeightSet weightSet : exerciseSession.getSets())
                     {
@@ -118,7 +118,7 @@ public class ExerciseGraphPresenter extends BasePresenter
                         }
                     }
                     // and finally construct our entry.
-                    returnList.add(new Entry(maxWeight, i, workoutSession));
+                    returnList.add(new Entry((float)maxWeight, i, workoutSession));
                 }
             }
         }
